Updated Windows project to be compilable as both static and dynamic lib.
Removed a few unneeded cpp files on the way.

Note: building as DLL produces many of the well known compiler warnings C4251 (type needs to have a DLL interface ..) because the runtime uses many STL classes in exported classes. I tried exporting all those types via explicit template instantion, but once I hit unordered_map I gave up. Just crazy what you have to export just to make this map export properly (and I already had like 50 exports already in place). So at the end I disabled this warning on project level. So make sure you build this DLL and all binaries using it with exact the same compiler and linker settings (same C++ runtime etc.).

+++ runtime/Cpp/runtime/CommonToken.cpp
@@ -128,7 +128,7 @@ void CommonToken::setCharPositionInLine(int charPositionInLine) {
   _charPositionInLine = charPositionInLine;
 }
 
-int CommonToken::getChannel() {
+size_t CommonToken::getChannel() {
   return _channel;
 }
